council-worker/requirements.txt
crewai>=0.28.0
anthropic>=0.24.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
supabase>=2.8.0
python-dotenv>=1.0.0
tavily-python>=0.3.0
httpx>=0.27.0
